backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["bakboem <ice.bakbum@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.13"
aio-pika = "9.5.4"
aiormq = "6.8.1"
annotated-types = "0.7.0"
anyio = "4.7.0"
asyncpg = "0.30.0"
bcrypt = "4.0.1"
certifi = "2024.12.14"
cffi = "1.17.1"
click = "8.1.8"
cryptography = "44.0.0"
dnspython = "2.7.0"
ecdsa = "0.19.0"
email-validator = "2.2.0"
exceptiongroup = "1.2.2"
fastapi = "^0.115.6"
fastapi-cli = "0.0.7"
greenlet = "3.1.1"
h11 = "0.14.0"
httpcore = "1.0.7"
httptools = "0.6.4"
httpx = "0.28.1"
idna = "3.10"
itsdangerous = "2.2.0"
jinja2 = "3.1.5"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
multidict = "6.1.0"
orjson = "3.10.12"
pamqp = "3.3.0"
passlib = "1.7.2"
propcache = "0.2.1"
psycopg2-binary = "2.9.10"
pyasn1 = "0.6.1"
pycparser = "2.22"
pydantic = "2.10.4"
pydantic-extra-types = "2.10.1"
pydantic-settings = "2.7.0"
pydantic-core = "2.27.2"
pygments = "2.18.0"
python-dotenv = "1.0.1"
python-jose = "3.3.0"
python-multipart = "0.0.20"
pytz = "2024.2"
pyyaml = "6.0.2"
rich = "13.9.4"
rich-toolkit = "0.12.0"
rsa = "4.9"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.36"
starlette = "0.41.3"
typer = "0.15.1"
typing-extensions = "4.12.2"
ujson = "5.10.0"
uvicorn = "0.34.0"
uvloop = "0.21.0"
watchfiles = "1.0.3"
websockets = "14.1"
yarl = "1.18.3"
colorlog = "^6.9.0"
redis = "^5.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
